feat: 기존의 주문확인 페이지를 활용하여 결제확인 페이지 구현 및 새로운 주문확인 페이지 생성
