+++ src/router.tsx
@@ -1,6 +1,7 @@
 import { createBrowserRouter } from "react-router";
 import ShoppingCartPage from "./pages/ShoppingCartPage/ShoppingCartPage";
 import OrderCheckPage from "./pages/OrderCheckPage/OrderCheckPage";
+import PaymentConfirmPage from "./pages/PaymentConfirmPage/PaymentConfirmPage";
 import Layout from "./Layout";
 
 export const routes = [
@@ -14,6 +15,11 @@ export const routes = [
       {
         path: "/order-check",
         element: <OrderCheckPage />,
+      },
+
+      {
+        path: "/payment-confirm",
+        element: <PaymentConfirmPage />,
       },
     ],
   },
